DeepForgeX/workshop/ms_dnn_model/esmm/requirements.txt
absl-py==2.2.2
aiobotocore==2.3.4
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioitertools==0.12.0
aiosignal==1.3.1
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.3.0
awscli==1.22.76
awscli-plugin-endpoint==0.4
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.1.0
boto3==1.21.21
botocore==1.24.21
cachetools==5.5.2
cattrs==22.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
clang==5.0
cloudpickle==2.1.0
colorama==0.4.3
comm==0.2.2
conda-pack @ file:///home/conda/feedstock_root/build_artifacts/conda-pack_1719323969738/work
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docutils==0.15.2
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.7.4
fastjsonschema==2.21.1
filelock==3.16.1
findspark==2.0.1
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2022.7.1
gast==0.3.3
gdown==5.2.0
google-auth==2.40.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
grpcio-tools==1.47.5
h11==0.16.0
h5py==2.10.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
metaspore==1.1.0
mistune==3.1.3
mmh3==3.0.0
multidict==6.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.9
notebook_shim==0.2.4
numpy==1.23.2
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.7.0
packaging==25.0
pandas==1.4.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9
pyarrow==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pymilvus==2.1.3
pymongo==4.2.0
PySocks==1.7.1
pyspark==3.1.2
python-consul==1.1.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==5.4.1
pyzmq==26.4.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.7.2
s3fs==2022.7.1
s3transfer==0.5.2
scikit-learn==1.3.2
scipy==1.5.0
Send2Trash==1.8.3
six==1.15.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.2.1
torch==1.11.0+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.4.0
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wrapt==1.12.1
yarl==1.15.2
zipp==3.20.2
